+++ HARDWARE/TIMER/timer.h
@@ -6,5 +6,5 @@ extern u8 UPDATE_CNT;
 extern u16 TIMER_CNT;
 extern u32 ELAPSE;
 
-void TIM2_Int_Init(u16 arr,u16 psc);
+void TIM2_Int_Init(u32 arr,u16 psc);
 #endif
